# docs/Enrollment_Platform_OpenAPI_Contract_Draft_v0.1.3.yaml
openapi: 3.1.2
...
info:
  title: Enrollment Platform - OpenAPI / Contract Draft v0.1.3
  version: 0.1.3
  summary: OpenAPI / Contract Draft v0.1.3 derived from Enrollment Protocol Specification v0.2.4
  description: 'Executable contract derived from Enrollment Protocol Specification v0.2.4. This revision incorporates CR-M5.6-001:
    deterministic originator-resource Replay Capsule AAD v1, exact same-secret recovery, original-result replay, and distinct
    permanent replay-unavailable (409) versus transient dependency (503) semantics; it does not add implementation scope.'
  x-document-status: DRAFT TECHNICAL CONTRACT - REVIEWED REVISION
  x-source-protocol: Enrollment Protocol Specification v0.2.4
  x-change-request: CR-M5.6-001
  x-open-008: CLOSED_AT_LOGICAL_PROTOCOL_LEVEL
servers:
- url: https://enrollment.example.invalid
  description: Placeholder public Reverse Proxy endpoint. Replace during deployment specification.
  x-placeholder: true
...
paths:
  /v1/me/authorizations:
    get:
      summary: List effective partner/scope authorizations for the authenticated principal
      operationId: getMyAuthorizations
      tags:
      - Authorization
      x-protocol-source: Enrollment Protocol Specification v0.2.4
      security:
      - HumanOIDC: []
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Effective current authorizations.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/XCorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyAuthorizationsResponse'
              example:
                principal_id: principal-123
                partners:
                - partner_id: P1
                  display_name: Revenda A
                  scopes:
                  - device:preonboard
                - partner_id: P2
                  display_name: Revenda B
                  scopes:
                  - device:preonboard
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
      x-error-codes:
        '401': &id001
        - AUTHENTICATION_REQUIRED
        '403': &id002
        - SCOPE_DENIED
        - PARTNER_NOT_AUTHORIZED
        '429': &id003
        - RATE_LIMITED
        '503': &id004
        - DEPENDENCY_UNAVAILABLE
  /v1/pre-onboarding-requests:
    post:
      summary: Create a pre-onboarding request
      operationId: createPreOnboardingRequest
      tags:
      - Pre-Onboarding
      description: Creates claimed/non-attested device proposal. partner_id is revalidated server-side. Temporary principal
        may submit but cannot approve. A successful same-scope/same-fingerprint replay returns the original committed 201
        representation, original ETag and original Location; X-Correlation-ID belongs to the current HTTP attempt.
      x-protocol-source: Enrollment Protocol Specification v0.2.4
      security:
      - HumanOIDC: []
      - TemporaryPrincipalToken: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreOnboardingCreateRequest'
            example:
              partner_id: P1
              claimed_device:
                hostname: PC-001
                serial_number: SN123
                smbios_uuid: '...'
                manufacturer: Example
                model: Workstation
                tpm_present: true
                tpm_vendor: '...'
                ek_public_hash: '...'
              agent:
                version: 1.0.0
                platform: windows
      responses:
        '201':
          description: Pre-onboarding request created. Request access token is returned once. A successful idempotent replay
            returns this original committed 201 representation, including the exact original request_access_token.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/XCorrelationId'
            ETag:
              $ref: '#/components/headers/ETag'
            Location:
              $ref: '#/components/headers/Location'
              description: 'Location of the created pre-onboarding request: /v1/pre-onboarding-requests/{pre_onboarding_request_id}.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreOnboardingCreateResponse'
              example:
                pre_onboarding_request_id: por-123
                status: PENDING_APPROVAL
                partner_id: P1
                expires_at: '2026-08-21T23:00:00Z'
                request_access_token: reqtok_ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789abcdef
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/IdempotencyReplayConflictOrUnavailable'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
      x-error-codes:
        '400': &id005
        - INVALID_REQUEST
        '401': *id001
        '403': *id002
        '409':
        - IDEMPOTENCY_CONFLICT
        - IDEMPOTENCY_REPLAY_UNAVAILABLE
        '413': &id007
        - PAYLOAD_TOO_LARGE
        '415': &id008
        - UNSUPPORTED_MEDIA_TYPE
        '429': *id003
        '503': *id004
      x-idempotent-secret-replay:
        secretField: request_access_token
        mechanism: encrypted-idempotency-replay-capsule
        sameSecretRequired: true
        activeVerifierStorage: non-reversible
        capsulePurpose: response-loss-recovery-only
        originator:
          operationId: createPreOnboardingRequest
          method: POST
          routeTemplate: /v1/pre-onboarding-requests
          resourceKind: PRE_ONBOARDING_REQUEST
          resourceId: pre_onboarding_request_id
          issuedCredentialType: REQUEST_ACCESS_TOKEN
          issuedCredentialVersion: 1
        effectiveScope: CredentialScope(credential kind, opaque non-secret server-authenticated credential binding) + canonical uppercase HTTP method + canonical OpenAPI route template + exact Idempotency-Key
        aadV1: [aad_domain, aad_version, originator_operation, originator_resource_kind, originator_resource_id, effective_scope_credential_kind, effective_scope_credential_binding, effective_scope_method, effective_scope_route_template, effective_scope_idempotency_key, request_fingerprint_version, request_fingerprint_digest, issued_credential_type, issued_credential_version]
        aadEncoding: deterministic, versioned, domain-separated, ordered, length-framed, unambiguous; integer/version fields are unsigned big-endian; no secret plaintext
        recovery: permanent loss of an authoritatively known committed result is 409 IDEMPOTENCY_REPLAY_UNAVAILABLE retryable=false; transient protector failure is 503 DEPENDENCY_UNAVAILABLE retryable=true; neither permits replacement minting
        replayAuthority: no mutation; no quota consumption; original result snapshot supplies body, ETag and Location; current-attempt correlation ID is excluded from scope, fingerprint and AAD
...
components:
  schemas:
    ResourceId:
      type: string
      minLength: 1
      maxLength: 4096
      description: Opaque server-issued resource identifier.
    Timestamp:
      type: string
      format: date-time
...
    PartnerAuthorization:
      type: object
      additionalProperties: true
      required:
      - partner_id
      - display_name
      - scopes
      properties:
        partner_id:
          $ref: '#/components/schemas/ResourceId'
        display_name:
          type: string
          minLength: 1
          maxLength: 4096
        scopes:
          type: array
          maxItems: 128
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 4096
      x-forward-compatible-response: true
    MyAuthorizationsResponse:
      type: object
      additionalProperties: true
      required:
      - principal_id
      - partners
      properties:
        principal_id:
          $ref: '#/components/schemas/ResourceId'
        partners:
          type: array
          maxItems: 128
          items:
            $ref: '#/components/schemas/PartnerAuthorization'
      x-forward-compatible-response: true
    ClaimedDevice:
      type: object
      additionalProperties: false
      required:
      - hostname
      description: Claimed/non-attested device attributes supplied before technical verification.
      properties:
        hostname:
          type: string
          minLength: 1
          maxLength: 4096
        serial_number:
          type:
          - string
          - 'null'
          maxLength: 4096
        smbios_uuid:
          type:
          - string
          - 'null'
          maxLength: 4096
        manufacturer:
          type:
          - string
          - 'null'
          maxLength: 4096
        model:
          type:
          - string
          - 'null'
          maxLength: 4096
        tpm_present:
          type:
          - boolean
          - 'null'
        tpm_vendor:
          type:
          - string
          - 'null'
          maxLength: 4096
        ek_public_hash:
          type:
          - string
          - 'null'
          maxLength: 4096
    AgentDescriptor:
      type: object
      additionalProperties: false
      required:
      - version
      - platform
      properties:
        version:
          type: string
          minLength: 1
          maxLength: 4096
        platform:
          type: string
          const: windows
    PreOnboardingCreateRequest:
      type: object
      additionalProperties: false
      required:
      - partner_id
      - claimed_device
      - agent
      properties:
        partner_id:
          $ref: '#/components/schemas/ResourceId'
        claimed_device:
          $ref: '#/components/schemas/ClaimedDevice'
        agent:
          $ref: '#/components/schemas/AgentDescriptor'
    PreOnboardingCreateResponse:
      type: object
      additionalProperties: true
      required:
      - pre_onboarding_request_id
      - status
      - partner_id
      - expires_at
      - request_access_token
      properties:
        pre_onboarding_request_id:
          $ref: '#/components/schemas/ResourceId'
        status:
          type: string
          const: PENDING_APPROVAL
        partner_id:
          $ref: '#/components/schemas/ResourceId'
        expires_at:
          $ref: '#/components/schemas/Timestamp'
        request_access_token:
          type: string
          minLength: 1
          maxLength: 4096
          description: Opaque server-generated secret. Active verifier is non-reversible. The SAME value may be reproduced
            only by idempotent replay through the temporary encrypted Replay Capsule during its valid window. Never log or
            place in URL.
      x-forward-compatible-response: true
...
    ProblemDetail:
      type: object
      required:
      - type
      - title
      - status
      - error_code
      - correlation_id
      - retryable
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
          maxLength: 4096
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type:
          - string
          - 'null'
          maxLength: 4096
        instance:
          type:
          - string
          - 'null'
          format: uri-reference
        error_code:
          type: string
          minLength: 1
          maxLength: 4096
        correlation_id:
          type: string
          minLength: 1
          maxLength: 4096
        retryable:
          type: boolean
        retry_after_seconds:
          type:
          - integer
          - 'null'
          minimum: 0
      additionalProperties: true
      description: RFC 9457 Problem Details plus stable domain extensions. Clients MUST use status/error_code, not human detail
        text, for logic.
...
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Printable ASCII idempotency key scoped by authenticated identity/credential scope + method + route template.
        16..128 characters.
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: ^[\x21-\x7E]{16,128}$
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: Optional caller correlation identifier; server generates one if absent.
      schema:
        type: string
        minLength: 1
        maxLength: 4096
...
  headers:
    XCorrelationId:
      description: End-to-end correlation identifier generated or validated by the platform.
      schema:
        type: string
        minLength: 1
        maxLength: 4096
    ETag:
      description: Strong HTTP entity-tag for a resource representation. Weak validators (W/) are not used for protocol concurrency.
      schema:
        type: string
        pattern: ^"[^\r\n]+"$
        maxLength: 4096
      example: '"4"'
    RetryAfter:
      description: Retry delay in seconds when applicable.
      schema:
        type: integer
        minimum: 0
    Location:
      description: URI-reference of the resource created or status resource associated with the operation.
      schema:
        type: string
        format: uri-reference
        maxLength: 4096
    WWWAuthenticate:
      description: HTTP authentication challenge. Required on bearer-authentication 401 responses; mTLS handshake failure
        may occur before an HTTP response exists.
      schema:
        type: string
        minLength: 1
        maxLength: 4096
      example: Bearer
  responses:
    BadRequest:
      description: Invalid request
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/XCorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: https://pki.example/errors/invalid-request
            title: Invalid request
            status: 400
            detail: Request cannot be completed under the current protocol state.
            instance: /v1/example
            error_code: INVALID_REQUEST
            correlation_id: 0198example
            retryable: false
    Unauthorized:
      description: Authentication required
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/XCorrelationId'
        WWW-Authenticate:
          $ref: '#/components/headers/WWWAuthenticate'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: https://pki.example/errors/authentication-required
            title: Authentication required
            status: 401
            detail: Request cannot be completed under the current protocol state.
            instance: /v1/example
            error_code: AUTHENTICATION_REQUIRED
            correlation_id: 0198example
            retryable: false
    Forbidden:
      description: Access denied
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/XCorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: https://pki.example/errors/scope-denied
            title: Access denied
            status: 403
            detail: Request cannot be completed under the current protocol state.
            instance: /v1/example
            error_code: SCOPE_DENIED
            correlation_id: 0198example
            retryable: false
...
    PayloadTooLarge:
      description: Payload too large
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/XCorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: https://pki.example/errors/payload-too-large
            title: Payload too large
            status: 413
            detail: Request cannot be completed under the current protocol state.
            instance: /v1/example
            error_code: PAYLOAD_TOO_LARGE
            correlation_id: 0198example
            retryable: false
    UnsupportedMediaType:
      description: Unsupported media type
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/XCorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: https://pki.example/errors/unsupported-media-type
            title: Unsupported media type
            status: 415
            detail: Request cannot be completed under the current protocol state.
            instance: /v1/example
            error_code: UNSUPPORTED_MEDIA_TYPE
            correlation_id: 0198example
            retryable: false
...
    TooManyRequests:
      description: Rate limited
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/XCorrelationId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: https://pki.example/errors/rate-limited
            title: Rate limited
            status: 429
            detail: Request cannot be completed under the current protocol state.
            instance: /v1/example
            error_code: RATE_LIMITED
            correlation_id: 0198example
            retryable: true
    ServiceUnavailable:
      description: Dependency unavailable
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/XCorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: https://pki.example/errors/dependency-unavailable
            title: Dependency unavailable
            status: 503
            detail: Request cannot be completed under the current protocol state.
            instance: /v1/example
            error_code: DEPENDENCY_UNAVAILABLE
            correlation_id: 0198example
            retryable: true
...
    IdempotencyReplayConflictOrUnavailable:
      description: 'Idempotency failure: IDEMPOTENCY_CONFLICT is same EffectiveScope/Idempotency-Key with a different request fingerprint; IDEMPOTENCY_REPLAY_UNAVAILABLE is permanent loss of recovery capability for an authoritatively known committed originator result. Both are 409 and retryable=false.'
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/XCorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: https://pki.example/errors/idempotency-replay-unavailable
            title: Idempotency replay unavailable
            status: 409
            detail: The committed originator result remains authoritative but its secret recovery material is permanently unavailable.
            instance: /v1/example
            error_code: IDEMPOTENCY_REPLAY_UNAVAILABLE
            correlation_id: 0198example
            retryable: false
...
  securitySchemes:
    HumanOIDC:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Human principal JWT issued/brokered by Keycloak. Agent login acquisition mechanism remains OPEN-003.
    AdminOIDC:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Administrative OIDC bearer JWT. Domain authorization and scopes are enforced by Enrollment API.
    TemporaryPrincipalToken:
      type: http
      scheme: bearer
      description: Short-lived contingency principal credential. Concrete Keycloak/bootstrap flow is outside this contract.
    RequestAccessToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque >=256-bit capability bound to one pre-onboarding request; single-use for INITIAL enrollment creation
        after approval. The originator response may reproduce the SAME token only by the protocol Idempotency Replay Capsule
        during the valid replay window; the active verifier remains non-reversible.
    EnrollmentAccessToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque >=256-bit bearer credential bound to one enrollment for status/evidence/delivery and installation-failure
        reporting. The originator response may reproduce the SAME token only by the protocol Idempotency Replay Capsule during
        the valid replay window; the active verifier remains non-reversible.
    DeviceMTLS:
      type: mutualTLS
      description: Device client certificate authentication terminates at the trusted Reverse Proxy. The Enrollment API server
        stub MUST NOT infer that it directly terminates the device TLS socket. Proxy-to-API uses separate service mTLS and
        protected certificate metadata.
      x-terminated-at: reverse-proxy
      x-upstream-authentication: proxy-service-mtls
      x-device-identity-propagation: protected-metadata-with-server-side-certificate-lookup
...